pyproject.toml
[tool.poetry]
name = "hermes"
version = "0.1.0"
description = "Slack Bolt App. Responds to file upload messages."
authors = ["Noland Seigler <57370691+nolandseigler@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
slack_bolt = "^1.7.0"
python-dotenv = "^0.18.0"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
pylint = "^2.9.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
